+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/extensions/CollectionExtensions.kt
@@ -27,11 +27,26 @@ infix fun <T> Array<out T>.anyIn(other: Array<out T>): Boolean = this.any { it i
 /**判断当前数组中的任意元素是否被另一集合包含。*/
 infix fun <T> Array<out T>.anyIn(other: Iterable<T>): Boolean = this.any { it in other }
 
+/**判断当前数组中的任意元素是否被另一序列包含。*/
+infix fun <T> Array<out T>.anyIn(other: Sequence<T>): Boolean = this.any { it in other }
+
 /**判断当前集合中的任意元素是否被另一数组包含。*/
 infix fun <T> Iterable<T>.anyIn(other: Array<out T>): Boolean = this.any { it in other }
 
 /**判断当前集合中的任意元素是否被另一集合包含。*/
 infix fun <T> Iterable<T>.anyIn(other: Iterable<T>): Boolean = this.any { it in other }
+
+/**判断当前集合中的任意元素是否被另一序列包含。*/
+infix fun <T> Iterable<T>.anyIn(other: Sequence<T>): Boolean = this.any { it in other }
+
+/**判断当前序列中的任意元素是否被另一数组包含。*/
+infix fun <T> Sequence<T>.anyIn(other: Array<out T>): Boolean = this.any { it in other }
+
+/**判断当前序列中的任意元素是否被另一集合包含。*/
+infix fun <T> Sequence<T>.anyIn(other: Iterable<T>): Boolean = this.any { it in other }
+
+/**判断当前序列中的任意元素是否被另一序列包含。*/
+infix fun <T> Sequence<T>.anyIn(other: Sequence<T>): Boolean = this.any { it in other }
 
 
 /**判断当前数组是否以指定元素开始。*/
@@ -55,24 +70,47 @@ infix fun <T> Iterable<T>.startsWith(elements: Array<out T>): Boolean = this.fir
 /**判断当前集合是否以指定元素结束。*/
 infix fun <T> Iterable<T>.endsWith(element: T): Boolean = this.firstOrNull() == element
 
-/**判断当前可迭代对象是否以任意指定元素结束。*/
+/**判断当前集合是否以任意指定元素结束。*/
 infix fun <T> Iterable<T>.endsWith(elements: Array<out T>): Boolean = this.firstOrNull() in elements
+
+/**判断当前序列是否以指定元素开始。*/
+infix fun <T> Sequence<T>.startsWith(element: T): Boolean = this.firstOrNull() == element
+
+/**判断当前序列是否以任意指定元素开始。*/
+infix fun <T> Sequence<T>.startsWith(elements: Array<out T>): Boolean = this.firstOrNull() in elements
+
+
+/**判断当前序列是否为空。*/
+fun <T> Sequence<T>.isEmpty() = !this.isNotEmpty()
+
+/**判断当前序列是否不为空。*/
+fun <T> Sequence<T>.isNotEmpty() = this.iterator().hasNext()
 
 
 /**如果当前数组不为空，则返回转换后的值。*/
 @Suppress("UPPER_BOUND_CANNOT_BE_ARRAY")
-inline fun <C : Array<*>> C.ifNotEmpty(transform: (C) -> C): C {
+inline fun <T : Array<*>> T.ifNotEmpty(transform: (T) -> T): T {
 	return if(this.isEmpty()) this else transform(this)
 }
 
-/**如果当前集合不为空，则返回转换后的值。推荐仅用于长链式方法调用。*/
-inline fun <C : Collection<*>> C.ifNotEmpty(transform: (C) -> C): C {
+/**如果当前集合不为空，则返回转换后的值。*/
+inline fun <T : Collection<*>> T.ifNotEmpty(transform: (T) -> T): T {
 	return if(this.isEmpty()) this else transform(this)
 }
 
-/**如果当前映射不为空，则返回转换后的值。推荐仅用于长链式方法调用。*/
-inline fun <C : Map<*, *>> C.ifNotEmpty(transform: (C) -> C): C {
+/**如果当前映射不为空，则返回转换后的值。*/
+inline fun <T : Map<*, *>> T.ifNotEmpty(transform: (T) -> T): T {
 	return if(this.isEmpty()) this else transform(this)
+}
+
+/**如果当前序列不为空，则返回重新生成的值。*/
+fun <T> Sequence<T>.ifNotEmpty(transform: () -> Sequence<T>): Sequence<T> = sequence {
+	val iterator = this@ifNotEmpty.iterator()
+	if(iterator.hasNext()) {
+		yieldAll(transform())
+	} else {
+		yieldAll(iterator)
+	}
 }
 
 
@@ -109,7 +147,6 @@ fun <T> MutableList<T>.removeAllAt(indices: IntRange) {
 	for(index in indices.reversed()) this.removeAt(index)
 }
 
-
 /**将指定索引的元素插入到另一索引处。后者为移动前的索引，而非移动后的索引。*/
 fun <T> MutableList<T>.move(fromIndices: Int, toIndex: Int): T {
 	val element = this[fromIndices]
@@ -141,7 +178,7 @@ inline fun <T, R : Any> Array<out T>.zipWithFirst(other: Array<out R>, predicate
 	return this.mapNotNull { e1 -> other.firstOrNull { e2 -> predicate(e1, e2) }?.let { e1 to it } }
 }
 
-/**绑定当前集合中的元素以及另一个集合中满足指定预测的首个元素。过滤总是不满足的情况。*/
+/**绑定当前数组中的元素以及另一个集合中满足指定预测的首个元素。过滤总是不满足的情况。*/
 inline fun <T, R : Any> Array<out T>.zipWithFirst(other: Iterable<R>, predicate: (T, R) -> Boolean): List<Pair<T, R>> {
 	return this.mapNotNull { e1 -> other.firstOrNull { e2 -> predicate(e1, e2) }?.let { e1 to it } }
 }
@@ -153,6 +190,11 @@ inline fun <T, R : Any> Iterable<T>.zipWithFirst(other: Array<out R>, predicate:
 
 /**绑定当前集合中的元素以及另一个集合中满足指定预测的首个元素。过滤总是不满足的情况。*/
 inline fun <T, R : Any> Iterable<T>.zipWithFirst(other: Iterable<R>, predicate: (T, R) -> Boolean): List<Pair<T, R>> {
+	return this.mapNotNull { e1 -> other.firstOrNull { e2 -> predicate(e1, e2) }?.let { e1 to it } }
+}
+
+/**绑定当前序列中的元素以及另一个序列中满足指定预测的首个元素。过滤总是不满足的情况。*/
+inline fun <T, R : Any> Sequence<T>.zipWithFirst(other: Sequence<R>, crossinline predicate: (T, R) -> Boolean): Sequence<Pair<T, R>> {
 	return this.mapNotNull { e1 -> other.firstOrNull { e2 -> predicate(e1, e2) }?.let { e1 to it } }
 }
 
@@ -180,55 +222,65 @@ private tailrec fun Map<String, Any?>.privateDeepGet(subPaths: List<String>): An
 		value is Array<*> -> value.toIndexKeyMap()
 		value is Iterable<*> -> value.toIndexKeyMap()
 		value is Map<*, *> -> value.toStringKeyMap()
+		value is Sequence<*> -> value.toIndexKeyMap()
 		else -> throw IllegalArgumentException("[ERROR] There is not a value related to this reference path.")
 	}
 	return fixedValue.privateDeepGet(currentSubPaths)
 }
 
 
-/**递归平滑映射当前数组，返回路径-值映射，默认使用标准引用[ReferenceCase.StandardReference]。可以指定层级[hierarchy]，默认为全部层级。*/
-fun <T> Array<out T>.deepFlatten(hierarchy: Int = -1, pathCase: ReferenceCase = ReferenceCase.StandardReference): Map<String, Any?> =
-	this.toIndexKeyMap().privateDeepFlatten(hierarchy, listOf(), pathCase)
+/**递归平滑映射当前数组，返回引用-值映射。默认使用标准引用[ReferenceCase.StandardReference]。可以指定层级，默认为全部层级。*/
+fun <T> Array<out T>.deepFlatten(hierarchy: Int = -1, pathFormatCase: FormatCase = ReferenceCase.StandardReference): Map<String, Any?> =
+	this.toIndexKeyMap().privateDeepFlatten(hierarchy, listOf(), pathFormatCase)
 
-/**递归平滑映射当前集合，返回路径-值映射，默认使用标准引用[ReferenceCase.StandardReference]。可以指定层级[hierarchy]，默认为全部层级。*/
-fun <T> Iterable<T>.deepFlatten(hierarchy: Int = -1, pathCase: ReferenceCase = ReferenceCase.StandardReference): Map<String, Any?> =
-	this.toIndexKeyMap().privateDeepFlatten(hierarchy, listOf(), pathCase)
+/**递归平滑映射当前集合，返回引用-值映射。默认使用标准引用[ReferenceCase.StandardReference]。可以指定层级，默认为全部层级。*/
+fun <T> Iterable<T>.deepFlatten(hierarchy: Int = -1, pathFormatCase: FormatCase = ReferenceCase.StandardReference): Map<String, Any?> =
+	this.toIndexKeyMap().privateDeepFlatten(hierarchy, listOf(), pathFormatCase)
 
-/**递归平滑映射当前映射，返回路径-值映射，默认使用标准引用[ReferenceCase.StandardReference]。可以指定层级[hierarchy]，默认为全部层级。*/
-fun <K, V> Map<K, V>.deepFlatten(hierarchy: Int = -1, pathCase: ReferenceCase = ReferenceCase.StandardReference): Map<String, Any?> =
-	this.toStringKeyMap().privateDeepFlatten(hierarchy, listOf(), pathCase)
+/**递归平滑映射当前映射，返回引用-值映射。默认使用标准引用[ReferenceCase.StandardReference]。可以指定层级，默认为全部层级。*/
+fun <K, V> Map<K, V>.deepFlatten(hierarchy: Int = -1, pathFormatCase: FormatCase = ReferenceCase.StandardReference): Map<String, Any?> =
+	this.toStringKeyMap().privateDeepFlatten(hierarchy, listOf(), pathFormatCase)
 
-private fun Map<String, Any?>.privateDeepFlatten(hierarchy: Int = -1, preSubPaths: List<String>, pathCase: ReferenceCase = ReferenceCase.StandardReference): Map<String, Any?> {
+/**递归平滑映射当前序列，返回引用-值映射。默认使用标准引用[ReferenceCase.StandardReference]。可以指定层级，默认为全部层级。*/
+fun <T> Sequence<T>.deepFlatten(hierarchy: Int = -1, pathFormatCase: FormatCase = ReferenceCase.StandardReference): Map<String, Any?> =
+	this.toIndexKeyMap().privateDeepFlatten(hierarchy, listOf(), pathFormatCase)
+
+private fun Map<String, Any?>.privateDeepFlatten(hierarchy: Int = -1, preSubPaths: List<String>, pathFormatCase: FormatCase = ReferenceCase.StandardReference): Map<String, Any?> {
 	return this.flatMap { (key, value) ->
 		val currentHierarchy = if(hierarchy <= 0) hierarchy else hierarchy - 1
 		//每次递归需要创建新的子路径列表
 		val currentPreSubPaths = preSubPaths + key
 		//如果不是集合类型，则拼接成完整路径，与值一同返回
 		val fixedValue = when {
-			currentHierarchy == 0 -> return@flatMap listOf(currentPreSubPaths.joinBy(pathCase) to value)
+			currentHierarchy == 0 -> return@flatMap listOf(currentPreSubPaths.joinBy(pathFormatCase) to value)
 			value is Array<*> -> value.toIndexKeyMap()
 			value is Iterable<*> -> value.toIndexKeyMap()
 			value is Map<*, *> -> value.toStringKeyMap()
-			else -> return@flatMap listOf(currentPreSubPaths.joinBy(pathCase) to value)
+			value is Sequence<*> -> value.toIndexKeyMap()
+			else -> return@flatMap listOf(currentPreSubPaths.joinBy(pathFormatCase) to value)
 		}
-		return@flatMap fixedValue.privateDeepFlatten(currentHierarchy, currentPreSubPaths, pathCase).toList()
+		return@flatMap fixedValue.privateDeepFlatten(currentHierarchy, currentPreSubPaths, pathFormatCase).toList()
 	}.toMap()
 }
 
 
-/**根据指定的Json路径[ReferenceCase.JsonSchemaReference]递归查询当前数组，返回匹配的路径-值映射，默认使用标准引用[ReferenceCase.StandardReference]*/
-fun <T> Array<out T>.deepQuery(path: String, pathCase: ReferenceCase = ReferenceCase.StandardReference): Map<String, Any?> =
-	this.toIndexKeyMap().privateDeepQuery(path.splitBy(ReferenceCase.JsonSchemaReference), listOf(), pathCase)
+/**根据指定的JsonSchema引用[ReferenceCase.JsonSchemaReference]递归查询当前数组，返回匹配的引用-值映射，默认使用标准引用[ReferenceCase.StandardReference]。*/
+fun <T> Array<out T>.deepQuery(path: String, referenceCase: ReferenceCase = ReferenceCase.StandardReference): Map<String, Any?> =
+	this.toIndexKeyMap().privateDeepQuery(path.splitBy(ReferenceCase.JsonSchemaReference), listOf(), referenceCase)
 
-/**根据指定的Json路径[ReferenceCase.JsonSchemaReference]递归查询当前集合，返回匹配的路径-值映射，默认使用标准引用[ReferenceCase.StandardReference]*/
-fun <T> Iterable<T>.deepQuery(path: String, pathCase: ReferenceCase = ReferenceCase.StandardReference): Map<String, Any?> =
-	this.toIndexKeyMap().privateDeepQuery(path.splitBy(ReferenceCase.JsonSchemaReference), listOf(), pathCase)
+/**根据指定的JsonSchema引用[ReferenceCase.JsonSchemaReference]递归查询当前集合，返回匹配的引用-值映射，默认使用标准引用[ReferenceCase.StandardReference]。*/
+fun <T> Iterable<T>.deepQuery(path: String, referenceCase: ReferenceCase = ReferenceCase.StandardReference): Map<String, Any?> =
+	this.toIndexKeyMap().privateDeepQuery(path.splitBy(ReferenceCase.JsonSchemaReference), listOf(), referenceCase)
 
-/**根据指定的Json路径[ReferenceCase.JsonSchemaReference]递归查询当前映射，返回匹配的路径-值映射，默认使用标准引用[ReferenceCase.StandardReference]*/
-fun <K, V> Map<K, V>.deepQuery(path: String, pathCase: ReferenceCase = ReferenceCase.StandardReference): Map<String, Any?> =
-	this.toStringKeyMap().privateDeepQuery(path.splitBy(ReferenceCase.JsonSchemaReference), listOf(), pathCase)
+/**根据指定的JsonSchema引用[ReferenceCase.JsonSchemaReference]递归查询当前映射，返回匹配的引用-值映射，默认使用标准引用[ReferenceCase.StandardReference]。*/
+fun <K, V> Map<K, V>.deepQuery(path: String, referenceCase: ReferenceCase = ReferenceCase.StandardReference): Map<String, Any?> =
+	this.toStringKeyMap().privateDeepQuery(path.splitBy(ReferenceCase.JsonSchemaReference), listOf(), referenceCase)
 
-private fun Map<String, Any?>.privateDeepQuery(subPaths: List<String>, preSubPaths: List<String>, pathCase: ReferenceCase = ReferenceCase.StandardReference): Map<String, Any?> {
+/**根据指定的JsonSchema引用[ReferenceCase.JsonSchemaReference]递归查询当前序列，返回匹配的引用-值映射，默认使用标准引用[ReferenceCase.StandardReference]。*/
+fun <T> Sequence<T>.deepQuery(path: String, referenceCase: ReferenceCase = ReferenceCase.StandardReference): Map<String, Any?> =
+	this.toIndexKeyMap().privateDeepQuery(path.splitBy(ReferenceCase.JsonSchemaReference), listOf(), referenceCase)
+
+private fun Map<String, Any?>.privateDeepQuery(subPaths: List<String>, preSubPaths: List<String>, referenceCase: ReferenceCase = ReferenceCase.StandardReference): Map<String, Any?> {
 	return this.flatMap { (key, value) ->
 		val currentSubPath = subPaths.first()
 		val currentSubPaths = subPaths.drop(1)
@@ -236,7 +288,8 @@ private fun Map<String, Any?>.privateDeepQuery(subPaths: List<String>, preSubPat
 		val currentPreSubPaths = preSubPaths + key
 		//如果不是集合类型，则拼接成完整路径，与值一同返回
 		val fixedValue = when {
-			currentSubPaths.isEmpty() -> return@flatMap listOf(currentPreSubPaths.joinBy(pathCase) to value).filter {
+			currentSubPaths.isEmpty() -> return@flatMap listOf(currentPreSubPaths.joinBy(referenceCase) to value).filter {
+				//TODO 提取为枚举
 				when {
 					//如果子路径表示一个列表或映射，例如："[]" "-" "{}"
 					currentSubPath in arrayOf("[]", "-", "{}") -> true
@@ -255,9 +308,10 @@ private fun Map<String, Any?>.privateDeepQuery(subPaths: List<String>, preSubPat
 			value is Array<*> -> value.toIndexKeyMap()
 			value is Iterable<*> -> value.toIndexKeyMap()
 			value is Map<*, *> -> value.toStringKeyMap()
+			value is Sequence<*> -> value.toIndexKeyMap()
 			else -> return@flatMap listOf<Pair<String, Any?>>()
 		}
-		return@flatMap fixedValue.privateDeepQuery(currentSubPaths, currentPreSubPaths, pathCase).toList()
+		return@flatMap fixedValue.privateDeepQuery(currentSubPaths, currentPreSubPaths, referenceCase).toList()
 	}.toMap()
 }
 
@@ -277,6 +331,11 @@ fun <K, V> Map<K, V>.toStringKeyMap(): Map<String, V> {
 	return this.mapKeys { (k, _) -> k.toString() }
 }
 
+/**将当前序列转化成以键为值的映射。*/
+fun <T> Sequence<T>.toIndexKeyMap(): Map<String, T> {
+	return this.withIndex().associate { (i, e) -> i.toString() to e }
+}
+
 
 /**得到指定索引的值，如果出错，则返回空字符串。*/
 fun Array<String>.getOrEmpty(index: Int): String = this.getOrElse(index) { "" }
@@ -286,39 +345,53 @@ fun List<String>.getOrEmpty(index: Int): String = this.getOrElse(index) { "" }
 
 
 /**去除第一行空白行。*/
-fun <C : CharSequence> Array<out C>.dropBlank(): List<C> = this.dropWhile { it.isBlank() }
+fun <T : CharSequence> Array<out T>.dropBlank(): List<T> = this.dropWhile { it.isBlank() }
 
 /**去除最后一行空白行。*/
-fun <C : CharSequence> Array<out C>.dropLastBlank(): List<C> = this.dropLastWhile { it.isBlank() }
+fun <T : CharSequence> Array<out T>.dropLastBlank(): List<T> = this.dropLastWhile { it.isBlank() }
 
 /**去除第一行空白行。*/
-fun <C : CharSequence> Iterable<C>.dropBlank(): List<C> = this.dropWhile { it.isBlank() }
+fun <T : CharSequence> Iterable<T>.dropBlank(): List<T> = this.dropWhile { it.isBlank() }
 
 /**去除最后一行空白行。*/
-fun <C : CharSequence> List<C>.dropLastBlank(): List<C> = this.dropLastWhile { it.isBlank() }
+fun <T : CharSequence> List<T>.dropLastBlank(): List<T> = this.dropLastWhile { it.isBlank() }
+
+/**去除第一行空白行。*/
+fun <T : CharSequence> Sequence<T>.dropBlank(): Sequence<T> = this.dropWhile { it.isBlank() }
 
 
 /**过滤空字符串。*/
-fun <C : CharSequence> Array<out C>.filterNotEmpty(): List<C> = this.filter { it.isNotEmpty() }
+fun <T : CharSequence> Array<out T>.filterNotEmpty(): List<T> = this.filter { it.isNotEmpty() }
 
 /**过滤空白字符串。*/
-fun <C : CharSequence> Array<out C>.filterNotBlank(): List<C> = this.filter { it.isNotEmpty() }
+fun <T : CharSequence> Array<out T>.filterNotBlank(): List<T> = this.filter { it.isNotEmpty() }
 
 /**过滤空字符串。*/
-fun <C : CharSequence> Iterable<C>.filterNotEmpty(): List<C> = this.filter { it.isNotEmpty() }
+fun <T : CharSequence> Iterable<T>.filterNotEmpty(): List<T> = this.filter { it.isNotEmpty() }
 
 /**过滤空白字符串。*/
-fun <C : CharSequence> Iterable<C>.filterNotBlank(): List<C> = this.filter { it.isNotEmpty() }
+fun <T : CharSequence> Iterable<T>.filterNotBlank(): List<T> = this.filter { it.isNotEmpty() }
+
+/**过滤空字符串。*/
+fun <T : CharSequence> Sequence<T>.filterNotEmpty(): Sequence<T> = this.filter { it.isNotEmpty() }
+
+/**过滤空白字符串。*/
+fun <T : CharSequence> Sequence<T>.filterNotBlank(): Sequence<T> = this.filter { it.isNotEmpty() }
 
 
-/**映射当前带索引值集合的索引，返回带有新的索引的带索引值集合。*/
-inline fun <T> Iterable<IndexedValue<T>>.mapIndices(transform: (Int) -> Int): Iterable<IndexedValue<T>> {
-	return this.map { (index, value) -> IndexedValue(transform(index), value) }
+/**将当前数组映射为转化索引后的索引-值对集合。*/
+inline fun <T> Array<T>.withIndex(transform: (Int) -> Int): Iterable<IndexedValue<T>> {
+	return this.withIndex().map { (i, v) -> IndexedValue(transform(i), v) }
 }
 
-/**映射当前带索引值集合的值，返回带有新的值的带索引值集合。*/
-inline fun <T> Iterable<IndexedValue<T>>.mapValues(transform: (T) -> T): Iterable<IndexedValue<T>> {
-	return this.map { (index, value) -> IndexedValue(index, transform(value)) }
+/**将当前集合映射为转化索引后的索引-值对集合。*/
+inline fun <T> Iterable<T>.withIndex(transform: (Int) -> Int): Iterable<IndexedValue<T>> {
+	return this.withIndex().map { (i, v) -> IndexedValue(transform(i), v) }
+}
+
+/**将当前序列映射为转化索引后的索引-值对序列。*/
+inline fun <T> Sequence<T>.withIndex(crossinline transform: (Int) -> Int): Sequence<IndexedValue<T>> {
+	return this.withIndex().map { (i, v) -> IndexedValue(transform(i), v) }
 }